abf/requirements.txt
aniso8601==7.0.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
easydict==1.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
graphviz==0.8.4
idna==2.8
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mxnet==1.5.0
networkx==2.3
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
urllib3==1.25.3
Werkzeug==1.0.0
